refactor : 반복문 stream 화, 미사용 코드 삭제
