+++ tripeer/src/main/java/com/j10d207/tripeer/place/db/entity/additional/AdditionalBaseEntity.java
@@ -3,8 +3,8 @@ package com.j10d207.tripeer.place.db.entity.additional;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.j10d207.tripeer.place.db.dto.AdditionalDto;
 
+import com.j10d207.tripeer.place.dto.res.AdditionalDto;
 import jakarta.persistence.MappedSuperclass;
 import jakarta.persistence.Id;
 import jakarta.persistence.GeneratedValue;
