+++ tripeer/src/main/java/com/j10d207/tripeer/place/dto/res/AdditionalDto.java
@@ -1,4 +1,4 @@
-package com.j10d207.tripeer.place.db.dto;
+package com.j10d207.tripeer.place.dto.res;
 
 import java.util.ArrayList;
 import java.util.List;
