--- tripeer/src/main/java/com/j10d207/tripeer/place/service/TownService.java
@@ -1,12 +1,0 @@
-package com.j10d207.tripeer.place.service;
-
-import com.j10d207.tripeer.place.dto.res.CityAndTownDTO;
-
-import java.util.List;
-
-public interface TownService {
-
-    public List<CityAndTownDTO.TownListDTO> searchTown(String cityId, String townName);
-    public CityAndTownDTO.TownListDTO townDetail(String townName);
-    public CityAndTownDTO getAllCityAndTown();
-}
